requirements.txt
gym
openai
selenium
Pillow
regex
PyYAML